third_party/WebKit/LayoutTests/bluetooth/service/getCharacteristics/gen-characteristic-not-found-with-uuid.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
bluetooth_test(() => {
  return getEmptyHealthThermometerService()
    .then(({service}) => assert_promise_rejects_with_message(
      service.getCharacteristics('battery_level'),
      new DOMException(
        'No Characteristics matching UUID ' + battery_level.uuid + ' found ' +
        'in Service with UUID ' + health_thermometer.uuid + '.',
        'NotFoundError')));
}, 'Request for absent characteristics with UUID. Reject with NotFoundError.');

</script>
